changed playSpeed data type to double in control bar

+++ MileStone1/MileStone1/controls/controlBar/IControlBarModel.cs
@@ -6,7 +6,7 @@ interface IControlBarModel : INotifyPropertyChanged
     bool running { get; set; }
     int lines { get; set; }
     int position { get; set; }
-    int playSpeed { get; set; }
+    double playSpeed { get; set; }
 
     void move();
 }
